requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
celery
redis
Pillow
psycopg2-binary
python-decouple